Add pytest basic settings and rename `_test` files that are not really tests to prevent incorrect discovery
