+++ requirements-dev.txt
@@ -1,2 +1,3 @@
 bumpversion>=0.5
+pytest
 -e .
