+++ src/compas/geometry/algorithms/_smoothing_cpp/smoothing_cpp_check_rhino.py
