+++ src/compas/numerical/algorithms/_fd_cpp/fd_cpp_check.py
